feat(inject): first steps towards xdotool inject fallback
