+++ espanso-inject/src/x11/xdotool/README.md
@@ -1,0 +1,2 @@
+This is a fallback injection module that relies on the awesome xdotool project: 
+https://github.com/jordansissel/xdotool
