+++ espanso/src/cli/worker/engine/dispatch/executor/event_injector.rs
@@ -67,6 +67,7 @@ impl<'a> TextInjector for EventInjectorAdapter<'a> {
         })
         .try_into()
         .unwrap(),
+      x11_use_xdotool_fallback: true, // TODO: put actual config
     };
 
     // We don't use the lines() method because it skips emtpy lines, which is not what we want.
